Interaction part
document.getElementById('animationProgress') is used to connect a range input (progress bar) in the HTML with the sketch.

animationProgress.value = sound.currentTime() continuously updates the bar as the music plays.
Added a click listener on the document to trigger music playback when the user clicks anywhere (document.addEventListener('click', ...)).
Time-Based Background Color Flash